add "refresh" context menu even if app server's down
Before: if the app server wasn't running when the Chrome extension was initialized, the parent context menu wouldn't be added at all, and the context menu couldn't be refreshed short of reloading the extension. After: always add the Shellac context menu with the "refresh" menu item, so app server can be brought up at any time.